pyproject.toml
[project]
name = "titania_xapi_client"
version = "0.0.1"
description = "Titania xAPI client"
authors = [
    { name = "Augusto Gonzalez", email = "agonzalez@tequila.capital" },
]
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Proprietary" }
keywords = ["trading", "opreations"]

dependencies = [
    "grpcio",
    "grpcio-tools",
    "pandas",
]

[project.urls]
Homepage = "https://github.com/tequila-capital/titania_xapi_client"
Repository = "https://github.com/tequila-capital/titania_xapi_client.git"
"Bug Tracker" = "https://github.com/tequila-capital/titania_xapi_client/issues"

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.package-data]
"titania_incubator" = ["data/*.json"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "--strict-markers --ignore=src"
testpaths = ["tests"]

[tool.ruff]
line-length = 99

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N"]
extend-select = ["E501"]

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["E402", "E501", "F401", "I001"]